GENERAL UPDATE
Cursor now locks when in game, and unlocks when in pause menu.
The Height of the Characater Controller has been decreased so the player
does not float above the ground.
The movement script is now complete. Moving in more than one direction
(i. e. diagonally) applies a 10% reduction in speed to avoid "snaking".
Torches are now gloomier.
MAKE SURE TO CHANGE THE PLAYER'S STATE USING THE FUNCTION. DO NOT ALTER
THE VARIABLE MANUALLY.